Add File menu (with save shortcut) to lab dialog (fix #1167)
